+++ src/utils/format-tokens.ts
@@ -1,0 +1,7 @@
+export function formatTokens(count: number): string {
+    if (count >= 1000000)
+        return `${(count / 1000000).toFixed(1)}M`;
+    if (count >= 1000)
+        return `${(count / 1000).toFixed(1)}k`;
+    return count.toString();
+}
